refactor: extract amount-calculation pipeline into DepositAmountHelper (#5007)
Phase 10d of #4631. Move the amount / balance / token↔fiat conversion
pipeline and the thin gasFeeHelper bridge shims out of
DepositFormViewModel into a new assisted-scope DepositAmountHelper:

- collectAmountChanges (+ lastTokenValueUserInput / lastFiatValueUserInput
  / amountChangesJob), updateTokenAmount and requireTokenAmount move
  verbatim, preserving the sequence-sensitive two-way binding semantics.
- getFeesFiatValue / convertAmountValue / calculateGasFee /
  getBitcoinTransactionPlan shims folded into the helper.
- DepositStrategyContext and DepositDataLoader.wireInitialState references
  re-pointed at the helper; viewModelScope ownership stays in the VM.

No public ViewModel API change.
